Documentation and New Metadata Package * Updated documentation for building a new package * Fixes to ConfigAdminAction * Refactor to move all AdminTab- modules to new `metadata` package   * ConfigMetadataModule   * WebActionMetadataModule * Move all AdminTab related install/multibinding to the AdminTabModule out of various other modules * Updated Ducks documentation * Redux DevTools Chrome Plugin * It's "T-Rex" not "TRex"
