+++ misk/src/main/kotlin/misk/web/metadata/ConfigMetadataModule.kt
@@ -1,15 +1,16 @@
-package misk.config
+package misk.web.metadata
 
+import misk.config.ConfigAdminAction
 import misk.environment.Environment
 import misk.inject.KAbstractModule
-import misk.web.actions.WebActionEntry
 import misk.web.actions.AdminTab
+import misk.web.actions.WebActionEntry
 import misk.web.proxy.WebProxyAction
 import misk.web.proxy.WebProxyEntry
 import misk.web.resources.StaticResourceAction
 import misk.web.resources.StaticResourceEntry
 
-class ConfigWebModule(val environment: Environment): KAbstractModule() {
+class ConfigMetadataModule(val environment: Environment) : KAbstractModule() {
   override fun configure() {
     multibind<WebActionEntry>().toInstance(WebActionEntry<ConfigAdminAction>())
     multibind<AdminTab>().toInstance(AdminTab(
