+++ misk/src/main/kotlin/misk/web/metadata/WebActionMetadataModule.kt
@@ -1,4 +1,4 @@
-package misk.web
+package misk.web.metadata
 
 import misk.environment.Environment
 import misk.inject.KAbstractModule
