+++ misk/web/tabs/loader/package.json
@@ -2,8 +2,6 @@
   "name": "misktab-loader",
   "version": "0.0.2",
   "main": "src/index.ts",
-  "repository": "git@github.com:square/misk.git",
-  "author": "Square/Misk Authors (https://github.com/square/misk/graphs/contributors)",
   "scripts": {
     "build": "cross-env NODE_ENV=production webpack",
     "clean": "rm -rf dist node_modules",
